Rename Selector to SearchSelector
