+++ src/components/SearchSelector/MultipleSearchSelector/styles.module.scss
