+++ src/components/SearchSelector/index.tsx
@@ -1,0 +1,1 @@
+export { SearchSelector } from "./SearchSelector";
