--- src/components/Selector/index.tsx
@@ -1,1 +1,0 @@
-export { Selector } from "./Selector";
